flask/requirements.txt
click==8.1.3
Flask==2.2.2
Flask-MySQLdb==1.0.1
gunicorn==20.1.0
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mysql-connector-python==8.0.32
mysqlclient==2.1.1
protobuf==3.20.3
Werkzeug==2.2.2
zipp==3.12.0
